requirements.txt
numpy
glfw
pyglm
PyOpenGL
PyOpenGL_accelerate
perlin-noise
pillow
pygltflib
open-cv-plus